Python files/requirements.txt
aiogram==3.20.0
SQLAlchemy==2.0.41
aiosqlite==0.21.0
python-dotenv==1.1.0
aiohttp==3.11.7
alembic==1.16.1